feat: fix mistakes and add more test to main

+++ fifth-lab/src/action/MilkAction.java
@@ -2,5 +2,5 @@ package action;
 
 public interface MilkAction extends ProductAction{
     public int getSize();
-    public int getFatPercentage();
+    public float getFatPercentage();
 }
